Show error messages generated by ‘curl’
The ‘-s’ switch makes ‘curl’ completely silent, even on error messages. So when
it fails it doesn't tell us why.

The ‘--no-progress-meter’ disables the normal progress meter output so that when
everything succeeds, there is no output, but when there is a failure we still
get to see what that is.